extract modulo_p and user_count_from_access_structure to Common module
